requirements.txt
streamlit
pandas
numpy
spacy
scikit-learn
python-docx
docx2txt
PyPDF2
pdfminer.six
nltk
